+++ zcash_client_memory/src/types/memory_wallet/mod.rs
@@ -518,8 +518,8 @@ impl<P: consensus::Parameters> MemoryWalletDb<P> {
     ) -> Result<Option<zip32::AccountId>, Error> {
         Ok(self
             .accounts
-            .iter()
-            .filter_map(|(_, a)| match a.source() {
+            .values()
+            .filter_map(|a| match a.source() {
                 AccountSource::Derived { derivation, .. } => {
                     if derivation.seed_fingerprint() == seed_fingerprint {
                         Some(derivation.account_index())
@@ -881,8 +881,8 @@ impl<P: consensus::Parameters> MemoryWalletDb<P> {
             // from the blocks table.
             let start_size = self
                 .accounts
-                .iter()
-                .filter_map(|(_, account)| {
+                .values()
+                .filter_map(|account| {
                     if account.birthday().height() == *birthday_height {
                         Some(account.birthday().sapling_frontier().tree_size())
                     } else {
@@ -959,8 +959,8 @@ impl<P: consensus::Parameters> MemoryWalletDb<P> {
             // from the blocks table.
             let start_size = self
                 .accounts
-                .iter()
-                .filter_map(|(_, account)| {
+                .values()
+                .filter_map(|account| {
                     if account.birthday().height() == *birthday_height {
                         Some(account.birthday().sapling_frontier().tree_size())
                     } else {