+++ zcash_primitives/src/transaction/tests.rs
@@ -150,15 +150,17 @@ fn zip_0143() {
     for tv in self::data::zip_0143::make_test_vectors() {
         let tx = Transaction::read(&tv.tx[..], tv.consensus_branch_id).unwrap();
         let signable_input = match tv.transparent_input {
-            Some(n) => {
-                SignableInput::Transparent(::transparent::sighash::SignableInput::from_parts(
+            Some(n) => SignableInput::Transparent(
+                ::transparent::sighash::SignableInput::from_parts(
+                    tx.transparent_bundle().unwrap(),
                     SighashType::parse(tv.hash_type as u8).unwrap(),
                     n as usize,
                     &tv.script_code,
                     &tv.script_code,
                     Zatoshis::from_nonnegative_i64(tv.amount).unwrap(),
-                ))
-            }
+                )
+                .unwrap(),
+            ),
             _ => SignableInput::Shielded,
         };
 
@@ -174,15 +176,17 @@ fn zip_0243() {
     for tv in self::data::zip_0243::make_test_vectors() {
         let tx = Transaction::read(&tv.tx[..], tv.consensus_branch_id).unwrap();
         let signable_input = match tv.transparent_input {
-            Some(n) => {
-                SignableInput::Transparent(::transparent::sighash::SignableInput::from_parts(
+            Some(n) => SignableInput::Transparent(
+                ::transparent::sighash::SignableInput::from_parts(
+                    tx.transparent_bundle().unwrap(),
                     SighashType::parse(tv.hash_type as u8).unwrap(),
                     n as usize,
                     &tv.script_code,
                     &tv.script_code,
                     Zatoshis::from_nonnegative_i64(tv.amount).unwrap(),
-                ))
-            }
+                )
+                .unwrap(),
+            ),
             _ => SignableInput::Shielded,
         };
 
@@ -332,13 +336,17 @@ fn zip_0244() {
             let value = bundle.authorization.input_amounts[index];
             let script_pubkey = &bundle.authorization.input_scriptpubkeys[index];
             let signable_input = |hash_type| {
-                SignableInput::Transparent(::transparent::sighash::SignableInput::from_parts(
-                    hash_type,
-                    index,
-                    script_pubkey,
-                    script_pubkey,
-                    value,
-                ))
+                SignableInput::Transparent(
+                    ::transparent::sighash::SignableInput::from_parts(
+                        bundle,
+                        hash_type,
+                        index,
+                        script_pubkey,
+                        script_pubkey,
+                        value,
+                    )
+                    .unwrap(),
+                )
             };
 
             assert_eq!(
@@ -434,7 +442,7 @@ fn zip_0233() {
         let input_scriptpubkeys = tv
             .script_pubkeys
             .iter()
-            .map(|s| Script(script::Code(s.clone())))
+            .map(|s| Script(script::Code(s.to_vec())))
             .collect();
 
         let test_bundle = txdata
@@ -479,8 +487,8 @@ fn zip_0233() {
         (tdata, txdata.digest(TxIdDigester))
     }
 
-    for tv in self::data::zip_0233::make_test_vectors() {
-        let (txdata, txid_parts) = to_test_txdata(&tv);
+    for tv in self::data::zip_0233::TEST_VECTORS {
+        let (txdata, txid_parts) = to_test_txdata(tv);
 
         assert_eq!(
             v6_signature_hash(&txdata, &SignableInput::Shielded, &txid_parts).as_ref(),