+++ zcash_proofs/CHANGELOG.md
@@ -10,9 +10,16 @@ workspace.
 
 ## [Unreleased]
 
+## [0.28.0] - 2026-06-02
+
+### Changed
+- Migrated to `zcash_primitives 0.28.0`.
+
+## [0.27.0] - 2026-04-27
+
 ### Changed
 - MSRV is now 1.85.1.
-- Migrated to `sapling-crypto 0.6`.
+- Migrated to `sapling-crypto 0.7`, `zcash_primitives 0.27`.
 
 ## [0.26.1] - 2025-10-18